No longer hide the default signature in main docs
The trick we were using to avoid displaying the ugly type
of the default method signature actually sucks, since it
makes writing manual instances of the class very awkward.

So instead we define type synonyms to capture each of the
requirements, which makes it easy to read and has the advantage
that it provides us with a natural place where to document
what types can have generic instances.