Temporarily kill adjProof
The current implementation of DictOf/ProofOf assumes Wear,
which made it possible to derive `adjProof` and `bproof`
for Barbie-types both with and without wear... but at the
moment this definition is of little use, since the `Wear`
bit is preventing us from doing anything useful.

We temporarily cripple 'ConstraintsB` so that we can still
get constraints for Wear-barbies while we think again how
to support both.