New version of ConstraintsB
This gives a definition that works for Wear/NonWear barbies and
and adjProof that works.

Essentially, for Wear-barbies we define the constraint for
`Wear f Int` as `(c (Wear f Int), Wear f Int ~ f Int)`. This way
we get a constraint that allows us to define instances for `Eq`,
etc.

Of course, this means that a) we need an overlapping instance for
`Eq Barbie Bare`, b) we will need to provide a different class
for the bare-constraints