Turn GenericN.Param into a tag
In the sense that it keeps, not only the position
of the type-param, but also it's original type.

This makes it easier to use since we can then write
patterns of the form:

```
instance GFoo (Rec (Param 0 f a) (f a)) ... where ...
instance Foo b => GFoo (Rec (b (Param 0 f)) (b f) ... where ...
instance GFoo (Rec a a) ... where
```
And we don't get overlapping instances.

Also, even if `Param` is exported, no harm is done, since
a user can't use `Param` in the definition of their types
in a way that would fool this type of instances