Deprecate Data.Barbie.Internal.Wear
The new instance deriving mechanism has made more or less clear
that the `Wear` hack is currently "magic", in the sense that the
usages of `unsafeCoerce` cannot be removed. This is bad, not
because it is unsafe (which I don't think it is), but because
users cannot write manual instances of even `FunctorB` for types
using `Wear`. We want the generic deriving to be a convenience,
here not a language extension.

The plan is to replace `Wear f a` by `Wear t f a`, where `t` is
a discriminator type that selects between applying and ignoring
the `f`. A "wear"-barbie, then becomes a type of the form `B t f`,
where we can derive, say, `instance FunctorB (B Covered)` using the
standard machinery.