Add new Wear implementation
It is based on a type discrimantor with two possible values,
`Bare` or `Covered`. So that we can have a wear-Barbie type
'B Covered f` that cannot be distinguished from a non-Wear
Barbie (except, perhaps, by the fact that `f` has a `nominal`
role because of the usage in the type family `Wear`).

Actually, it is a bit annoying that `f` gets a `nominal` role
given that it is actually used in a `representional` way inside
the definition of `Wear`. This may one day be fixed, see
[this ghc ticket]( https://ghc.haskell.org/trac/ghc/ticket/8177).