Can derive ConstraintsB for parametric barbies
Changes the way we generically compute AllB, so
we no longer use a closed type-family for the Rec case.

We need to be explicit about whether an occurrence of
`b f` corresponds to a recursive usage of `b` or not,
otherwise ghc will blow the stack trying to compute
the constraints.  So we add a type-function on `RepN (b f)`
that tags recursive usages so we can break recursion,
as before.

With this implementation we no longer need to pass `b`
as an extra parameter of all classes, which cleans up
the code a bit (although compile times seem to go up,
due to the extra type-family computation, I think).