Replace ProofB by ProductBC
It is easier to explain this way what this class is
and how it can be useful. Ideally, I'd would have put
`buniqC` as the sole method and implement `bdicts` in
terms of it, but ghc seems to choke when the type
signature of `buniqC` is used as a class method and
tries to use a default implementation of the method.