Derive more generic instances
Because the applicative laws are more lax then product ones,
we can derive new ApplicativeB instances: before we wanted
every field in a record to be covered by the argument functor,
while now it is also ok if we have some Monoid instead.

While adding tests for this, we noticed that for parametric
barbies, we were not able to derive `TraversableB` and
`ConstraintsB` if the parameter was being used naked in
a field, which should be perfectly valid as well.