Mark only the relevant parameter for gen deriv
We don't really need to annotate all the parameters with their parameter
number to derive most of the classes. In fact, doing this was
complicating most cases, since we were forced to consider cases where
a type was an irrelevant parameter or not, etc.

In fact, once we found yet another corner case that we were not
covering, it was better to just simplify things this way.

Notice that for GAllB we do want to mark all parameters, this has the
side-effect of letting `TagSelf` identify the recursive case when we
have parametric types. To be fair, I still don't quite see why ghc gets
stuck evaluating `TagSelf 0 (ParB b) (b X)`, but well...